refactor: 페이지 개수 데이터 타입 수정

+++ perfume-api/src/main/java/io/perfume/api/brand/adapter/in/http/dto/GetMagazineRequestDto.java
@@ -2,13 +2,13 @@ package io.perfume.api.brand.adapter.in.http.dto;
 
 import io.perfume.api.brand.application.port.in.dto.GetMagazineCommand;
 
-public record GetMagazineRequestDto(Long pageSize, String before, String after) {
+public record GetMagazineRequestDto(Integer pageSize, String before, String after) {
 
   public GetMagazineCommand toCommand(long brandId) {
     return new GetMagazineCommand(getSizeOrDefault(), before, after, brandId);
   }
 
-  private Long getSizeOrDefault() {
+  private int getSizeOrDefault() {
     return pageSize == null ? 10 : pageSize;
   }
 }
