+++ perfume-api/src/main/java/io/perfume/api/review/application/facade/dto/ReviewCommentDetailCommand.java
@@ -2,7 +2,7 @@ package io.perfume.api.review.application.facade.dto;
 
 import io.perfume.api.review.application.in.dto.GetReviewCommentsCommand;
 
-public record ReviewCommentDetailCommand(long reviewId, long size, String after, String before) {
+public record ReviewCommentDetailCommand(long reviewId, int size, String after, String before) {
 
   public GetReviewCommentsCommand toGetReviewCommentsCommand() {
     return new GetReviewCommentsCommand(size, before, after, reviewId);
